+++ vitest.config.ts
@@ -1,11 +1,11 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
     include: [
-      'packages/**/lib/**/*.spec.ts',
-      'packages/**/src/**/*.spec.ts',
-      'tests/**/*.spec.ts',
+      'packages/**/lib/**/*.spec.{ts,tsx}',
+      'packages/**/src/**/*.spec.{ts,tsx}',
+      'tests/**/*.spec.{ts,tsx}',
     ],
     environment: 'node',
     globals: false,
@@ -13,4 +13,4 @@ export default defineConfig({
       provider: 'v8',
     },
   },
-});
+})